pyproject.toml
[tool.poetry]
name = "wb-api"
version = "0.3.0"
description = "Библиотека для удобной работы с API Wildberries"
authors = ["artemdorozhkin"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.6,<4.0"
pydantic = ">=2.8.2"
requests = ">=2.32.3"



[tool.poetry.group.dev.dependencies]
python-dotenv = ">=1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
